D3D10 renderer continue

+++ src/priv/gfx/rndr/ogl/oglDrawMgr.cpp
@@ -164,7 +164,7 @@ void GN::gfx::OGLRenderer::drawEnd()
 //
 // -----------------------------------------------------------------------------
 void GN::gfx::OGLRenderer::clearScreen(
-    const GN::Vector4f & c, float z, UInt32 s, BitFields flags )
+    const GN::Vector4f & c, float z, UInt8 s, BitFields flags )
 {
     GN_GUARD_SLOW;
