+++ src/priv/gfx/rndr/ogl/oglRenderer.h
@@ -303,7 +303,7 @@ namespace GN { namespace gfx
 
         virtual bool drawBegin();
         virtual void drawEnd();
-        virtual void clearScreen( const Vector4f & c, float z, UInt32 s, BitFields flags );
+        virtual void clearScreen( const Vector4f & c, float z, UInt8 s, BitFields flags );
         virtual void drawIndexed( PrimitiveType prim,
                                   size_t        numprim,
                                   size_t        startvtx,
